Added EmbraceHttpClient abstraction
